requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
